Implement conversion of profiles 2.1->2.2
This works more or less the same as for machine instances. The code was copied from version 2.1 of the unserialise function of profiles, and optimised a bit. The output function is written from scratch. It has some code duplication. Maybe we have to do something about that.

Contributes to issue CURA-844.